+++ tools/extraChart/tool.js
@@ -990,6 +990,8 @@
       const fill = (container, count, color) => {
         let elem = $('div.lichessTools-goodMoves', container);
         if (!elem.length) {
+          const adviceSummary = $('div.advice-summary__player', container);
+          if (adviceSummary.length) {
           elem = $('<div></div>')
             .addClass('lichessTools-goodMoves')
             .addClass('advice-summary__error')
@@ -1023,11 +1025,14 @@
               this.safeSetActiveElements(chart,elems,dataset);
               chart.update('none');
             })
-            .insertAfter($('div.advice-summary__player', container));
+            .insertAfter(adviceSummary);
+          }
         }
-        elem.toggleClassSafe('symbol', !!count);
-        const newText = (count || 0).toString();
-        $('strong', elem).replaceText(newText);
+        if (elem.length) {
+          elem.toggleClassSafe('symbol', !!count);
+          const newText = (count || 0).toString();
+          elem.find('strong').replaceText(newText);
+        }
       };
       let container = $('div.advice-summary__side').get(0);
       const count = arr.filter(n => {